requirements.txt
# === Core Framework ===
streamlit>=1.39.0
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0

# === Utilities ===
python-dotenv>=1.0.0

# === Web Search ===
tavily-python>=0.5.0

# === Data / SQL (Layer 3) ===
sqlalchemy>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0

# === RAG (Layer 4) ===
chromadb>=0.5.0
pypdf>=4.0.0
langchain-text-splitters>=0.3.0

# === MCP (Layer 5) ===
mcp>=1.0.0
fastmcp>=0.4.0

# === Observability (Layer 9) ===
langsmith>=0.1.0
plotly>=5.18.0

# === Deployment (Layer 10) ===
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.9
